+++ CHANGELOG.md
@@ -1,2 +1,15 @@
+# Changes in 0.4.0
+
+* Threaded terrain collision generation
+* Threaded world loading
+* Smoke rendering performance improvements
+* Small memory allocation and performance improvements for chunk generation
+
+# Changes in 0.3.1
+
+* Fixed multiple crashes caused by logging out and back in into any world
+* Tiny performance improvement for comfort level calculation
+
 # Changes in 0.3.0
+
 * First public release